+++ pkg/chain/gen/contract.go
@@ -9,7 +9,6 @@ import (
 	"io/ioutil"
 	"os"
 	"path"
-	"regexp"
 	"strings"
 	"text/template"
 
@@ -17,84 +16,25 @@ import (
 	"golang.org/x/tools/imports"
 )
 
-// The extracted name + payability of methods from ABI JSON.
-type methodPayableInfo struct {
-	Name    string
-	Payable bool
-}
-
-var (
-	classNameRegexp *regexp.Regexp
-	shortVarRegexp  *regexp.Regexp
-)
-
-func init() {
-	var err error
-	classNameRegexp, err = regexp.Compile("ImplV.*")
-	if err != nil {
-		panic(fmt.Sprintf(
-			"Failed to compile class name regular expression: [%v].",
-			"ImplV.*",
-		))
-	}
-
-	shortVarRegexp, err = regexp.Compile("([A-Z])[^A-Z]*")
-	if err != nil {
-		panic(fmt.Sprintf(
-			"Failed to compile class name regular expression: [%v].",
-			"([A-Z])[^A-Z]*",
-		))
-	}
-}
-
-// The following structs are sent into the templates for compilation.
-type contractInfo struct {
-	Class           string
-	AbiClass        string
-	FullVar         string
-	ShortVar        string
-	ConstMethods    []methodInfo
-	NonConstMethods []methodInfo
-	Events          []eventInfo
-}
-
-type methodInfo struct {
-	CapsName          string
-	LowerName         string
-	Payable           bool
-	Params            string
-	ParamDeclarations string
-	Return            returnInfo
-}
-
-type returnInfo struct {
-	Multi        bool
-	Type         string
-	Declarations string
-	Vars         string
-}
-
-type eventInfo struct {
-	CapsName                  string
-	LowerName                 string
-	IndexedFilters            string
-	ParamExtractors           string
-	ParamDeclarations         string
-	IndexedFilterDeclarations string
-}
-
-// Main function. Expect <executable> [input.abi] [output.go] .
+// Main function. Expects to be invoked as:
+//
+//   <executable> [input.abi] contract/[contract_output.go]
+//
+// The first file will receive a contract binding that is slightly higher-level
+// than abigen's output, including an event-based interface for contract event
+// interaction, support for revert error reporting, serialized transaction
+// submission, and simplified transactor handling.
 func main() {
 	if len(os.Args) != 3 {
 		panic(fmt.Sprintf(
-			"Expected `%v [input.abi] [output.go]`, but got [%v].",
+			"Expected `%v [input.abi] [contract_output.go]`, but got [%v].",
 			os.Args[0],
 			os.Args,
 		))
 	}
 
 	abiPath := os.Args[1]
-	outputPath := os.Args[2]
+	contractOutputPath := os.Args[2]
 
 	abiFile, err := ioutil.ReadFile(abiPath)
 	if err != nil {
@@ -135,182 +75,60 @@ func main() {
 	abiClassName = abiClassName[0 : len(abiClassName)-4] // strip .abi
 	contractInfo := buildContractInfo(abiClassName, &abi, payableInfo)
 
-	buf, err := generateCode(outputPath, templates, &contractInfo)
+	contractBuf, err := generateCode(
+		contractOutputPath,
+		templates,
+		"contract.go.tmpl",
+		&contractInfo,
+	)
 	if err != nil {
 		panic(fmt.Sprintf(
-			"Failed to generate Go file at [%v]: [%v].",
-			outputPath,
+			"Failed to generate Go file for contract [%v] at [%v]: [%v].",
+			contractInfo.AbiClass,
+			contractOutputPath,
 			err,
 		))
 	}
 
-	// Save the promise code to a file.
-	if err := saveBufferToFile(buf, outputPath); err != nil {
+	// Save the contract code to a file.
+	if err := saveBufferToFile(contractBuf, contractOutputPath); err != nil {
 		panic(fmt.Sprintf(
 			"Failed to save Go file at [%v]: [%v].",
-			outputPath,
+			contractOutputPath,
 			err,
 		))
 	}
+
 }
 
-func buildContractInfo(
-	abiClassName string,
-	abi *abi.ABI,
-	payableInfo []methodPayableInfo,
-) contractInfo {
-	payableMethods := make(map[string]struct{})
-	for _, methodPayableInfo := range payableInfo {
-		if methodPayableInfo.Payable {
-			payableMethods[methodPayableInfo.Name] = struct{}{}
-		}
-	}
+// Generates code by applying the named template in the passed template bundle
+// to the specified data object. Writes the output to a buffer and then
+// formats and organizes the imports on that buffer, returning the final result
+// ready for emission onto the filesystem.
+//
+// Note that this means the generated file must compile, or import organization
+// will fail. The error message in case of compilation failure will be bubbled
+// up, but the file contents currently will not be written.
+func generateCode(
+	outFile string,
+	templat *template.Template,
+	templateName string,
+	data interface{},
+) (*bytes.Buffer, error) {
+	var buffer bytes.Buffer
 
-	goClassName := classNameRegexp.ReplaceAll([]byte(abiClassName), nil)
-	shortVar := strings.ToLower(string(shortVarRegexp.ReplaceAll(
-		[]byte(goClassName),
-		[]byte("$1"),
-	)))
-	constMethods, nonConstMethods := buildMethodInfo(payableMethods, abi.Methods)
-	events := buildEventInfo(abi.Events)
-
-	return contractInfo{
-		string(goClassName),
-		abiClassName,
-		lowercaseFirst(string(goClassName)),
-		string(shortVar),
-		constMethods,
-		nonConstMethods,
-		events,
-	}
-}
-
-func buildMethodInfo(
-	payableMethods map[string]struct{},
-	methodsByName map[string]abi.Method,
-) (constMethods []methodInfo, nonConstMethods []methodInfo) {
-	nonConstMethods = make([]methodInfo, 0, len(methodsByName))
-	constMethods = make([]methodInfo, 0, len(methodsByName))
-
-	for name, method := range methodsByName {
-		_, payable := payableMethods[name]
-		paramDeclarations := ""
-		params := ""
-
-		for index, param := range method.Inputs {
-			goType := param.Type.Type.String()
-			paramName := param.Name
-			if paramName == "" {
-				paramName = fmt.Sprintf("arg%v", index)
-			}
-
-			paramDeclarations += fmt.Sprintf("%v %v,\n", paramName, goType)
-			params += fmt.Sprintf("%v,\n", paramName)
-		}
-
-		returned := returnInfo{}
-		if len(method.Outputs) > 1 {
-			returned.Multi = true
-			returned.Type = strings.Replace(name, "get", "", 1)
-
-			for _, output := range method.Outputs {
-				goType := output.Type.Type.String()
-
-				returned.Declarations += fmt.Sprintf(
-					"\t%v %v\n",
-					uppercaseFirst(output.Name),
-					goType,
-				)
-				returned.Vars += fmt.Sprintf("%v,", output.Name)
-			}
-		} else if len(method.Outputs) == 0 {
-			returned.Multi = false
-		} else {
-			returned.Multi = false
-			returned.Type = method.Outputs[0].Type.Type.String()
-			returned.Vars += "ret,"
-		}
-
-		info := methodInfo{
-			uppercaseFirst(name),
-			lowercaseFirst(name),
-			payable,
-			params,
-			paramDeclarations,
-			returned,
-		}
-
-		if method.Const {
-			constMethods = append(constMethods, info)
-		} else {
-			nonConstMethods = append(nonConstMethods, info)
-		}
-	}
-
-	return constMethods, nonConstMethods
-}
-
-func buildEventInfo(eventsByName map[string]abi.Event) []eventInfo {
-	eventInfos := make([]eventInfo, 0, len(eventsByName))
-	for name, event := range eventsByName {
-		paramDeclarations := ""
-		paramExtractors := ""
-		indexedFilterDeclarations := ""
-		indexedFilters := ""
-		for _, param := range event.Inputs {
-			upperParam := uppercaseFirst(param.Name)
-			goType := param.Type.Type.String()
-
-			paramDeclarations += fmt.Sprintf("%v %v,\n", upperParam, goType)
-			paramExtractors += fmt.Sprintf("event.%v,\n", upperParam)
-			if param.Indexed {
-				indexedFilterDeclarations += fmt.Sprintf("%vFilter []%v,\n", param.Name, goType)
-				indexedFilters += fmt.Sprintf("%vFilter,\n", param.Name)
-			}
-		}
-
-		paramDeclarations += "blockNumber uint64,\n"
-		paramExtractors += "event.Raw.BlockNumber,\n"
-
-		eventInfos = append(eventInfos, eventInfo{
-			uppercaseFirst(name),
-			lowercaseFirst(name),
-			indexedFilters,
-			paramExtractors,
-			paramDeclarations,
-			indexedFilterDeclarations,
-		})
-	}
-
-	return eventInfos
-}
-
-func uppercaseFirst(str string) string {
-	return strings.ToUpper(str[0:1]) + str[1:]
-}
-
-func lowercaseFirst(str string) string {
-	return strings.ToLower(str[0:1]) + str[1:]
-}
-
-// Generates a code from template and configuration.
-// Returns a buffered code.
-func generateCode(outFile string, tmpl *template.Template, info *contractInfo) (*bytes.Buffer, error) {
-	var buf bytes.Buffer
-
-	if err := tmpl.ExecuteTemplate(&buf, "contract.go.tmpl", info); err != nil {
+	if err := templat.ExecuteTemplate(&buffer, templateName, data); err != nil {
 		return nil, fmt.Errorf(
-			"generating code for contract [%v] failed: [%v]",
-			info.AbiClass,
+			"generating code failed: [%v]",
 			err,
 		)
 	}
 
-	if err := organizeImports(outFile, &buf); err != nil {
+	if err := organizeImports(outFile, &buffer); err != nil {
 		return nil, err
 	}
 
-	return &buf, nil
+	return &buffer, nil
 }
 
 // Resolves imports in a code stored in a Buffer.
